Clk: Enable i2c1 to manager i2c1 clk rate during sys runtime
update as below:
Some clk is disabled and unprepared, but clk rate is not 0.
For example, i2c1 is disabled in dts file, but its status is
enabled and unprepared, so rate/voltage will be initialized by DVFS.
This will impact vdd-core's voltage.

So we enable i2c1 module to manager clk rate.

Bug 200493082

Change-Id: I1515edd2734a454820cbfafa272dcd1a9c900938
Signed-off-by: Taylor Xiao <txiao@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2033985
Reviewed-by: Terry Wang <terwang@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vijay Mali <vmali@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>