DT: t201b01: mdarcy 2551: Enable i2c1 to manager clk rate
Update as below:
Some clk is disabled and unprepared, but clk rate is not 0.
For example, i2c1 is disabled in dts file, but its status is
enabled and unprepared, so rate/voltage will be initialized by DVFS.
This will impact vdd-core's voltage.

So we enable i2c1 module to manager clk rate.

Bug 200493082

Change-Id: Ic9f636232d2f0072d98560c52e8088ed7ff295bd
Signed-off-by: Taylor Xiao <txiao@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2073685
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Robert Shih <rshih@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: David Dastous St Hilaire <ddastoussthi@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>